moved services to a new project. solution refactoring, reorganizing unit test. separating api and service tests.
