+++ ApplicationTracker.ServicesTests/LocationServiceTests.cs
@@ -1,26 +1,24 @@
 ﻿using ApplicationTracker.Data.Entities;
-using ApplicationTracker.Services;
-using ApplicationTrackerTests.Helpers;
+using ApplicationTracker.TestUtilities.Helpers;
 using Microsoft.Extensions.Logging;
 using Moq;
-using NUnit.Framework.Internal;
 
-namespace ApplicationTrackerTests.Services
+namespace ApplicationTracker.Services.Tests
 {
     [TestFixture]
-    public class JobTitlesServiceTests
+    public class LocationServiceTests
     {
-        private Mock<ILogger<JobTitleService>> _mockLogger;
-        private JobTitleService _service;
+        private Mock<ILogger<LocationService>> _mockLogger;
+        private LocationService _service;
 
         [SetUp]
         public void Setup()
         {
             // create a in memery context with 4 rows.
-            var context = ContextHelper.GetInMemoryContext<JobTitle>(4);
-            
-            _mockLogger = new Mock<ILogger<JobTitleService>>();
-            _service = new JobTitleService(context, _mockLogger.Object);
+            var context = ContextHelper.GetInMemoryContext<Location>(4);
+
+            _mockLogger = new Mock<ILogger<LocationService>>();
+            _service = new LocationService(context, _mockLogger.Object);
         }
 
         [Test]
@@ -28,7 +26,7 @@ namespace ApplicationTrackerTests.Services
         {
             // Setup
             var testId = 1;
-            var expected = $"Test {typeof(JobTitle).Name} {testId}";
+            var expected = $"Test {typeof(Location).Name} {testId}";
 
             // Act
             var result = await _service.GetAllAsync();
