+++ ApplicationTracker.ServicesTests/SourcesServiceTests.cs
@@ -1,25 +1,25 @@
 ﻿using ApplicationTracker.Data.Entities;
 using ApplicationTracker.Services;
-using ApplicationTrackerTests.Helpers;
+using ApplicationTracker.TestUtilities.Helpers;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace ApplicationTrackerTests.Services
+namespace ApplicationTracker.Services.Tests
 {
     [TestFixture]
-    public class WorkEnvironmentServiceTests
+    public class SourcesServiceTests
     {
-        private Mock<ILogger<WorkEnvironmentService>> _mockLogger;
-        private WorkEnvironmentService _service;
+        private Mock<ILogger<SourceService>> _mockLogger;
+        private SourceService _service;
 
         [SetUp]
         public void Setup()
         {
             // create a in memery context with 4 rows.
-            var context = ContextHelper.GetInMemoryContext<WorkEnvironment>(4);
+            var context = ContextHelper.GetInMemoryContext<Source>(4);
 
-            _mockLogger = new Mock<ILogger<WorkEnvironmentService>>();
-            _service = new WorkEnvironmentService(context, _mockLogger.Object);
+            _mockLogger = new Mock<ILogger<SourceService>>();
+            _service = new SourceService(context, _mockLogger.Object);
         }
 
         [Test]
@@ -27,13 +27,13 @@ namespace ApplicationTrackerTests.Services
         {
             // Setup
             var testId = 1;
-            var expected = $"Test {typeof(WorkEnvironment).Name} {testId}";
+            var expected = $"Test {typeof(Source).Name} {testId}";
 
             // Act
             var result = await _service.GetAllAsync();
 
             // Assert
-            
+
             Assert.Multiple(() =>
             {
                 Assert.That(result, Is.Not.Null);
