+++ ApplicationTracker.ServicesTests/WorkEnvironmentServiceTests.cs
@@ -1,26 +1,25 @@
 ﻿using ApplicationTracker.Data.Entities;
 using ApplicationTracker.Services;
-using ApplicationTrackerTests.Helpers;
+using ApplicationTracker.TestUtilities.Helpers;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace ApplicationTrackerTests.Services
+namespace ApplicationTracker.Services.Tests
 {
     [TestFixture]
-    public class OrganizationServiceTests
+    public class WorkEnvironmentServiceTests
     {
-        private Mock<ILogger<OrganizationService>> _mockLogger;
-        private OrganizationService _service;
+        private Mock<ILogger<WorkEnvironmentService>> _mockLogger;
+        private WorkEnvironmentService _service;
 
         [SetUp]
         public void Setup()
         {
             // create a in memery context with 4 rows.
-            var context = ContextHelper.GetInMemoryContext<Organization>(4);
+            var context = ContextHelper.GetInMemoryContext<WorkEnvironment>(4);
 
-            _mockLogger = new Mock<ILogger<OrganizationService>>();
-            _service = new OrganizationService(context, _mockLogger.Object);
-
+            _mockLogger = new Mock<ILogger<WorkEnvironmentService>>();
+            _service = new WorkEnvironmentService(context, _mockLogger.Object);
         }
 
         [Test]
@@ -28,12 +27,13 @@ namespace ApplicationTrackerTests.Services
         {
             // Setup
             var testId = 1;
-            var expected = $"Test {typeof(Organization).Name} {testId}";
+            var expected = $"Test {typeof(WorkEnvironment).Name} {testId}";
 
             // Act
             var result = await _service.GetAllAsync();
 
             // Assert
+
             Assert.Multiple(() =>
             {
                 Assert.That(result, Is.Not.Null);
