+++ ApplicationTrackerTests/Controllers/OrganizationsControllerTests.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework.Internal;
 
-namespace ApplicationTrackerTests.Controllers
+namespace ApplicationTracker.Tests.Controllers
 {
     [TestFixture]
     public class OrganizationsControllerTests
@@ -39,7 +39,7 @@ namespace ApplicationTrackerTests.Controllers
             var mockServiceProvider = new Mock<IServiceProvider>();
             mockServiceProvider
                 .Setup(sp => sp.GetService(typeof(OrganizationService)))
-                .Returns(_mockService.Object); 
+                .Returns(_mockService.Object);
 
             _serviceFactory = new ServiceFactory(mockServiceProvider.Object);
             _controller = new OrganizationsController(_serviceFactory, _mockLogger.Object);
@@ -60,7 +60,7 @@ namespace ApplicationTrackerTests.Controllers
             Assert.Multiple(() =>
             {
                 Assert.That(result.Result, Is.InstanceOf<OkObjectResult>());
-            
+
                 var okResult = result.Result as OkObjectResult;
                 Assert.That(okResult, Is.Not.Null);
 
@@ -86,7 +86,7 @@ namespace ApplicationTrackerTests.Controllers
             Assert.Multiple(() =>
             {
                 Assert.That(result.Result, Is.InstanceOf<NotFoundObjectResult>());
-                
+
                 var notFoundResult = result.Result as NotFoundObjectResult;
                 Assert.That(notFoundResult, Is.Not.Null);
 
@@ -117,10 +117,10 @@ namespace ApplicationTrackerTests.Controllers
             Assert.Multiple(() =>
             {
                 Assert.That(result.Result, Is.InstanceOf<OkObjectResult>());
-            
+
                 var okResult = result.Result as OkObjectResult;
                 Assert.That(okResult, Is.Not.Null);
-            
+
                 var returnedOrganization = okResult!.Value as OrganizationDto;
                 Assert.That(returnedOrganization, Is.Not.Null);
                 Assert.That(returnedOrganization!.Id, Is.EqualTo(testId));
@@ -144,10 +144,10 @@ namespace ApplicationTrackerTests.Controllers
             Assert.Multiple(() =>
             {
                 Assert.That(result.Result, Is.InstanceOf<NotFoundObjectResult>());
-            
+
                 var notFoundResult = result.Result as NotFoundObjectResult;
                 Assert.That(notFoundResult, Is.Not.Null);
-            
+
                 var errorResponse = notFoundResult!.Value as ErrorResponse;
                 Assert.That(errorResponse, Is.Not.Null);
                 Assert.That(errorResponse!.Message, Is.EqualTo("Organization not found"));
